Updated up to 25 June 2009, but not all 25 June files. From ndattani's scienide2 folder:
/Backup_Scienide2_2017_06_05/programs/betaFIT

It is bizarre though that a betaf.x was created, because the Makefile of 25 June 2009 doesn't make betaf.x anymore, it makes betafRet.x.

297991 Jun 25  2009 betafRet.x   [not included, because betaf.x is added in this commit]
 41212 Jun 25  2009 llsqf.o      [ignoring .o files]
 97164 Jun 25  2009 nllssrr.o    [ignoring .o files]
   936 Jun 25  2009 Makefile     [added]
146328 Jun 25  2009 betaf.x      [added]
 65755 Jun 25  2009 betaFITret.f [added as betafit.f]
  8663 Jun 13  2009 AF3X3.f      [added]
  7725 Jun 13  2009 AF3x3workingButWrong.f [added to other_files/Li2_historcal]
 42792 Jun 13  2009 debug3x3.x               "
  8790 Jun 13  2009 debug3x3.f               "

Ok here we go:

297991 2009-06-25 22:00:52 -0400 betafRet.x
 41212 2009-06-25 22:00:52 -0400 llsqf.o
 97164 2009-06-25 22:00:52 -0400 nllssrr.o
   936 2009-06-25 21:26:07 -0400 Makefile
146328 2009-06-25 21:24:39 -0400 betaf.x
 65755 2009-06-25 20:55:21 -0400 betaFITret.f

so I must have run the old Makefile on betaFITret.f not knowing that the betaf.x produced would still use the old betafit.f (but I'm still keeping this betaf.x because who knows what other changes were made that were written over and therefore not appearing in the commit history).
so then I changed the Makefile and recompiled to get betafRet.x

WARNING, betaf.x was probably NOT created with betaFITret.f, it was probably created with the most recent betafit.f before it, as you can see the Makefile was modified 2 minutes after that betaf.x was created, seems like very concrete evidence that the Makefile just had the old .f in it and now has the right .f, leading to the betafRet.x which we're going to commit in the next commit.